+++ packages/docs/build.config.ts
@@ -1,10 +1,16 @@
 import { defineBuildConfig } from 'unbuild'
 
+import pkg from './package.json'
+
 export default defineBuildConfig({
   // If entries is not provided, will be automatically inferred from package.json
   entries: [
-    // default
-    './index',
+    // node
+    './node/index',
+
+    // types
+    './types/index',
+
     // mkdist builder transpiles file-to-file keeping original sources structure
     // {
     //   builder: 'mkdist',
@@ -14,4 +20,12 @@ export default defineBuildConfig({
   ],
 
   declaration: true,
+
+  externals: [
+    ...Object.keys(pkg.devDependencies || {}),
+
+    'unocss/vite',
+    'unplugin-vue-components/vite',
+    '@shikijs/vitepress-twoslash/cache-fs',
+  ],
 })