+++ packages/docs/node/config/shims.ts
@@ -1,0 +1,3 @@
+import { dirname } from 'node:path'
+
+export const _dirname = dirname(new URL(import.meta.url).pathname)