+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## 0.0.1 (2025-02-21)
+## 0.0.3 (2025-03-20)
 
 
 ### Features
 
+* **docs:** packages/docs config & components ([78acea7](https://github.com/YunYouJun/docs/commit/78acea785ca4f4ccdfd841c6e0f7b4a3a3cfac6b))
+* export @yunyoujun/docs package ([c2452c7](https://github.com/YunYouJun/docs/commit/c2452c7cc4c47822e8745f3864c6786b6d38bda5))
 * init basic vitepress structure ([f62f41b](https://github.com/YunYouJun/docs/commit/f62f41b0e9faeb82de950d153c013667e966b610))